refactor(Badge): split files
